fix excp_level setting

+++ hwtvar.h
@@ -85,7 +85,7 @@ int hwt_process_create(int *sockpair, char **cmd, char **env, int *pid0);
 int hwt_process_start(int *sockpair);
 int hwt_record_fetch(struct trace_context *tc, int *nrecords);
 void hwt_procexit(pid_t pid, int status);
-size_t hwt_get_offs(struct trace_context *tc, size_t *offs);
+int hwt_get_offs(struct trace_context *tc, size_t *offs);
 void hwt_sleep(void);
 int hwt_elf_count_libs(const char *elf_path, uint32_t *nlibs0);
 int hwt_find_sym(struct trace_context *tc);
